refactor: update new import css module name
